+++ src/stac_asset/http_client.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+from asyncio import Queue
 from types import TracebackType
-from typing import AsyncIterator, Optional, Type, TypeVar
+from typing import Any, AsyncIterator, Optional, Type, TypeVar
 
 from aiohttp import ClientSession
 from yarl import URL
@@ -9,6 +10,7 @@ from yarl import URL
 from . import validate
 from .client import Client
 from .config import Config
+from .messages import OpenUrl
 
 T = TypeVar("T", bound="HttpClient")
 
@@ -36,13 +38,17 @@ class HttpClient(Client):
         self.check_content_type = check_content_type
 
     async def open_url(
-        self, url: URL, content_type: Optional[str] = None
+        self,
+        url: URL,
+        content_type: Optional[str] = None,
+        queue: Optional[Queue[Any]] = None,
     ) -> AsyncIterator[bytes]:
         """Opens a url with this client's session and iterates over its bytes.
 
         Args:
             url: The url to open
             content_type: The expected content type
+            queue: An optional queue to use for progress reporting
 
         Yields:
             AsyncIterator[bytes]: An iterator over the file's bytes
@@ -56,6 +62,8 @@ class HttpClient(Client):
                 validate.content_type(
                     actual=response.content_type, expected=content_type
                 )
+            if queue:
+                await queue.put(OpenUrl(url=url, size=response.content_length))
             async for chunk, _ in response.content.iter_chunks():
                 yield chunk