+++ tests/test_cli.py
@@ -24,7 +24,7 @@ def test_download_item_stdin_stdout(tmp_path: Path, item: Item) -> None:
         item_as_str = json.dumps(
             item.to_dict(include_self_link=True, transform_hrefs=False)
         )
-        runner = CliRunner()
+        runner = CliRunner(mix_stderr=False)
         result = runner.invoke(stac_asset._cli.cli, ["download"], input=item_as_str)
         assert result.exit_code == 0, result.stdout
         Item.from_dict(json.loads(result.stdout))
@@ -41,7 +41,7 @@ def test_download_item_collection_stdin_stdout(
         item_collection_as_str = json.dumps(
             item_collection.to_dict(transform_hrefs=False)
         )
-        runner = CliRunner()
+        runner = CliRunner(mix_stderr=False)
         result = runner.invoke(
             stac_asset._cli.cli, ["download"], input=item_collection_as_str
         )